Moved merge tags testing
